+++ Hippocampus/server/index.js
@@ -1,47 +1,14 @@
 const express = require('express')
-const mysql = require('mysql')
 const cors = require('cors')
-const { response } = require('express')
+const routes = require('./routes')
 
 const app = express()
 const port = 3001
 app.use(express.json())
 app.use(cors())
 
-const db = mysql.createConnection({
-    user: "root",
-    host: "localhost",
-    password: "secret",
-    database: "hippocampus",
-})
+app.post('/register', routes.register)
 
-app.post('/register', (req, res) =>{
-
-    const userEmail = req.body.userEmail
-    const username = req.body.username
-    const password = req.body.password
-
-    db.query("INSERT INTO users (username, password, email) VALUES (?, ?, ?)", [username, password, userEmail],
-    (err, result) => {
-        console.log(err)
-    })
-})
-
-app.post('/login', (req, res) =>{
-    const userEmail = req.body.userEmail
-    const password = req.body.password
-
-    db.query("SELECT * FROM users WHERE email = ? AND password = ?", [userEmail, password],
-    (err, result) => {
-        if(err){
-            res.send({err: err})
-        }
-            if(result.length > 0){
-                res.send(result)
-            }else{
-                res.send({ message: "Wrong email or password!" })
-            }
-    })
-})
+app.post('/login', routes.login)
 
 app.listen(port, () => console.log('Hippocampus listening on port ' + port + '!'))
